+++ inst/tinytest/test-pkg-glmmTMB.R
@@ -1,4 +1,6 @@
 source("helpers.R")
+exit_file("glmmTMB produces weird SEs")
+
 # if (!EXPENSIVE) exit_file("EXPENSIVE")
 using("marginaleffects")
 # exit_file("glmmTMB always causes problems")